refactor(preset): refactor function functions
